pyproject.toml
[tool.poetry]
name = "job-flow"
version = "0.1.0"
description = "A job-matching platform backend using FastAPI"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
firebase-admin = "^6.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
openai = "^1.6.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
pypdf2 = "^3.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
